+++ app/controllers/profiles_controller.rb
@@ -1,35 +1,9 @@
 class ProfilesController < ApplicationController
+  include MypageTurboActions
   before_action :set_user, only: %i[edit update]
 
   def show
     @myposts = current_user.posts.where(is_draft: false).order(created_at: :desc).page(params[:page])
-  end
-
-  def my_post_order
-    posts = current_user.posts.where(is_draft: false).order(created_at: :desc).page(params[:page])
-    render turbo_stream: turbo_stream.replace(
-      "js-my-post-order",
-      partial: 'profiles/profile_order',
-      locals: { posts: posts, my_post: '自分の投稿' }
-    )   
-  end
-
-  def my_draft_order
-    posts = current_user.posts.where(is_draft: true).order(created_at: :desc).page(params[:page])
-    render turbo_stream: turbo_stream.replace(
-      "js-my-post-order",
-      partial: 'profiles/profile_order',
-      locals: { posts: posts, my_post: '下書き' }
-    )   
-  end
-  
-  def my_reiatu_order
-    posts = current_user.reiatu_posts.order(created_at: :desc).page(params[:page])
-    render turbo_stream: turbo_stream.replace(
-      "js-my-post-order",
-      partial: 'profiles/profile_order',
-      locals: { posts: posts, my_post: 'いいねした投稿' }
-    )   
   end
 
   def edit;end
